# dymer-webserver/swagger_webserver.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "DYMER API",
    "description": "Documentation of <b>Webserver</b> module."
  },
  "host": "localhost:0000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/dservice/api/v1/opn/configs": {
      "get": {
        "tags": [
          "Services"
        ],
        "description": "",
        "parameters": [
          {
            "name": "req",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/dservice/api/v1/fwadapter/configs": {
      "get": {
        "tags": [
          "Services"
        ],
        "description": "",
        "parameters": [
          {
            "name": "req",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/dservice/api/v1/authconfig/": {
      "get": {
        "tags": [
          "Services"
        ],
        "description": "",
        "parameters": [
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Services"
        ],
        "description": "",
        "parameters": [
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/dservice/api/v1/authconfig/{id}": {
      "put": {
        "tags": [
          "Services"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Services"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/userinfo": {
      "post": {
        "tags": [
          "Webserver"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}